fix after a code review
