+++ src/utils/random-user.tsx
@@ -1,11 +1,11 @@
 const baseUrl = 'https://randomuser.me/api/?seed=motivation&inc=picture,name';
 
-export const getUsers = async (
-  count: number = 20,
-  page: number = 1,
-  callback: (users: any) => void
-) => {
-  const users = await fetch(`${baseUrl}&page=${page}&results=${count}`);
-  const parsedUsers = await users.json();
-  callback(parsedUsers.results);
+export const getUsers = async (count: number = 20, page: number = 1) => {
+  try {
+    const users = await fetch(`${baseUrl}&page=${page}&results=${count}`);
+    const parsedUsers = await users.json();
+    return parsedUsers.results;
+  } catch (error) {
+    alert(error);
+  }
 };
